requirements.txt
captcha
Pillow
numpy
pytest